ship editor done

+++ Assets/Scripts/Ship/Entities/ShipConstants.cs
@@ -5,6 +5,6 @@ public class ShipConstants
 {
     public static float SHIP_PART_SIZE = 1f;
 
-    public static float SHIP_GRID_WIDTH = 5f;
-    public static float SHIP_GRID_HEIGHT = 3f;
+    public static int SHIP_GRID_WIDTH = 5;
+    public static int SHIP_GRID_HEIGHT = 3;
 }
